Add CSS modules; Restructure folders; Refactor components, add pages; Styling in-progress
