+++ client/src/main.jsx
@@ -3,7 +3,6 @@ import ReactDOM from "react-dom/client";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 
 import App from "./App";
-import "./index.css";
 
 const queryClient = new QueryClient();
