+++ client/src/pages/NoteKeeperPage.jsx
@@ -1,0 +1,14 @@
+import { NotesPanel } from "../components/main/NotesPanel";
+
+export const NoteKeeperPage = () => {
+  /* Description of the layout
+   <Header /> -- TODO, navbar on top with search bar etc
+   <NotesPanel /> -- TODO main notes functionality
+   <Sidebar /> - TODO, 'container' for shortcuts like trashcan, reminders etc
+   */
+  return (
+    <>
+      <NotesPanel />
+    </>
+  );
+};
