+++ client/src/styles/Notes.module.css
@@ -1,0 +1,6 @@
+.notesGrid {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  grid-template-rows: auto;
+  gap: 25px;
+}
