+++ client/src/styles/NotesPanel.module.css
@@ -1,0 +1,6 @@
+.notesPanel {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 50px;
+}
